refactor(header): extract NavLink component to separate file
Move the NavLink component from header.tsx to a dedicated nav-link
.tsx file. This improves code organization and modularity, making the
 component easier to maintain and test. The refactoring also ensures
 cleaner imports and separation of concerns within the header module.
